MAC Clean-up [Issue-ID : ODUHIGH-212]

+++ src/5gnrmac/mac_upr_inf_api.h
@@ -19,12 +19,9 @@
 /* This file contains the definitions for Upper Interface APIs that are
  * invoked from MAC */
 
-#include "du_app_mac_inf.h"
-
-
-uint16_t MacDuAppSlotInd(Pst *pst, SlotInfo *slotInfo);
-uint16_t MacDuAppStopInd(Pst *pst, MacCellStopInfo *cellStopId);
-uint16_t MacDuAppUlCcchInd(Pst *pst, UlCcchIndInfo *ulCcchIndInfo);
+uint8_t MacDuAppSlotInd(Pst *pst, SlotInfo *slotInfo);
+uint8_t MacDuAppStopInd(Pst *pst, MacCellStopInfo *cellStopId);
+uint8_t MacDuAppUlCcchInd(Pst *pst, UlCcchIndInfo *ulCcchIndInfo);
 /**********************************************************************
          End of file
 **********************************************************************/
